requirements.txt
pandas>=2.0
yfinance>=0.2.31
jinja2>=3.1
pyyaml>=6.0
